% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addOne.R
\name{addOne}
\alias{addOne}
\alias{addOne_arma}
\alias{addOne_rcpp}
\title{Example addOne}
\usage{
addOne_arma(x)

addOne_rcpp(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A vector or univariate matrix.
}
\description{
Example addOne
}
\examples{

addOne_arma(1:3)

addOne_rcpp(1:3)
}
