% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knnDemo.R
\name{knnDemo}
\alias{knnDemo}
\title{k-Nearest-Neighbors Search}
\usage{
knnDemo(reference, query = NULL, k, verbose = FALSE)
}
\arguments{
\item{reference, query}{Matrix for query and reference points.}

\item{k}{An integer. Number of nearest neighbours to find.}

\item{verbose}{Should a log info be displayed? Default is \code{FALSE}.}
}
\value{
A \code{list} with two elements: \code{distances} and \code{neighbors}.
}
\description{
k-Nearest-Neighbors Search
}
\examples{
set.seed(1234)
x <- matrix(rnorm(10*5), ncol = 5)

res <- knnDemo(x, k = 3)

# mlpack's knn
res2 <- mlpack::knn(query = x, reference = x, k = 3)

# sanity check
all.equal(res, res2[-3]) # TRUE
}
