% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlpack.Rpkg-package.R
\docType{package}
\name{mlpack.Rpkg-package}
\alias{mlpack.Rpkg}
\alias{mlpack.Rpkg-package}
\title{mlpack.Rpkg: Demo Package using 'mlpack' header via 'Rcpp'}
\description{
Provides a demo R function 'knnDemo()' that sends its input to a custom C++ routine which then invokes the mlpack's functionality.
}
\keyword{internal}
